% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{correlation_coefficient}
\alias{correlation_coefficient}
\title{Compute the correlation coefficient}
\usage{
correlation_coefficient(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric value representing the correlation coefficient
}
\description{
This function computes the correlation coefficient between two vectors.
}
\examples{
correlation_coefficient(c(2.5, 3, 3.5, 2), c(1, 1.5, 2, 2.5))

}
