% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{create_sum_variable}
\alias{create_sum_variable}
\title{Create a new variable as a sum of two other variables}
\usage{
create_sum_variable(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric vector representing the sum of x and y
}
\description{
This function creates a new variable that is the sum of two existing variables.
}
\examples{
create_sum_variable(c(2.5, 3, 3.5, 2), c(1, 1.5, 2, 2.5))

}
