% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{multiple_regression_analysis}
\alias{multiple_regression_analysis}
\title{Perform multiple linear regression analysis}
\usage{
multiple_regression_analysis(formula, data)
}
\arguments{
\item{formula}{a formula specifying the regression model}

\item{data}{a data frame containing the variables specified in the formula}
}
\value{
a list containing the results of the regression analysis
}
\description{
This function performs a multiple linear regression analysis on the specified data and formula.
}
\examples{
data <- data.frame(x1 = c(1, 2, 3, 4), x2 = c(2, 3, 4, 5), x3 = c(3, 4, 5, 6), y = c(5, 6, 7, 8))
multiple_regression_analysis(formula = y ~ x1 + x2 + x3, data = data)

}
