% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create a scatter plot of two variables}
\usage{
scatter_plot(x, y)
}
\arguments{
\item{x}{a numeric vector representing the first variable}

\item{y}{a numeric vector representing the second variable}
}
\value{
a ggplot object representing the scatter plot
}
\description{
This function creates a scatter plot of two variables
}
\examples{
scatter_plot(c(2.5, 3, 3.5, 2), c(1, 1.5, 2, 2.5))
}
