% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{t_test_p_value}
\alias{t_test_p_value}
\title{Compute the p-value of a t-test}
\usage{
t_test_p_value(x, y)
}
\arguments{
\item{x}{a numeric vector representing the first group of data}

\item{y}{a numeric vector representing the second group of data}
}
\value{
a numeric value representing the p-value of the t-test
}
\description{
This function computes the p-value of a t-test between two vectors.
}
\examples{
t_test_p_value(c(2.5, 3, 3.5, 2), c(1, 1.5, 2, 2.5))
}
